backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis and Task Queue
redis==5.0.1
celery==5.3.6

# Rate Limiting
slowapi==0.1.9

# Web Scraping
selenium==4.17.2
beautifulsoup4==4.12.3
lxml==5.1.0
webdriver-manager==4.0.1

# AI Integration
openai==1.10.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
